`renewal-configuration/new-enrollment?enrollment_factory_id={ID}` is now `enrollment-factory/{ID}/onboard` updating some table constraints and indexes
