+++ examples/staging_demo/_testrun.txt
@@ -80,11 +80,9 @@ According to the above, `acme_server_id = 2; PEBBLE TEST 2`::
 
     ssl_manage data_testrun renewal-configuration list
 
-    ssl_manage data_testrun renewal-configuration new-enrollment \
-        enrollment_factory_id=1 \
+    ssl_manage data_testrun enrollment-factory onboard \
+        id=1 \
         domain_name="a.peter-sslers.testing.opensource.aptise.com" \
-        label="chall_prefix-{DOMAIN}" \
-        is_export_filesystem=enrollment_factory_default
 
 ## Run the audit tools
