+++ src/peter_sslers/lib/db/update.py
@@ -796,12 +796,12 @@ def update_EnrollmentFactory(
     private_key_cycle__backup: Optional[str],
     private_key_technology__backup: Optional[str],
     acme_profile__backup: Optional[str],
-    name: Optional[str],
+    name: str,
     note: Optional[str],
     domain_template_http01: Optional[str],
     domain_template_dns01: Optional[str],
-    label_template: Optional[str],
-    is_export_filesystem_id: Optional[int],
+    label_template: str,
+    is_export_filesystem_id: int,
 ) -> bool:
     if not any(
         (
@@ -813,18 +813,17 @@ def update_EnrollmentFactory(
         raise errors.InvalidTransition("Missing Required Primary.")
 
     # these require some validation
-    name = lib_utils.normalize_unique_text(name) if name else None
-    if name:
-        if name.startswith("rc-") or name.startswith("global"):
-            raise ValueError("`name` contains a reserved prefix or is a reserved word")
+    name = lib_utils.normalize_unique_text(name)
+    if name.startswith("rc-") or name.startswith("global"):
+        raise ValueError("`name` contains a reserved prefix or is a reserved word")
 
-        existingEnrollmentFactory = get__EnrollmentFactory__by_name(ctx, name)
-        if existingEnrollmentFactory and (
-            existingEnrollmentFactory.id != dbEnrollmentFactory.id
-        ):
-            raise errors.InvalidTransition(
-                "An EnrollmentFactory already exists with this name."
-            )
+    existingEnrollmentFactory = get__EnrollmentFactory__by_name(ctx, name)
+    if existingEnrollmentFactory and (
+        existingEnrollmentFactory.id != dbEnrollmentFactory.id
+    ):
+        raise errors.InvalidTransition(
+            "An EnrollmentFactory already exists with this name."
+        )
 
     # default to original
     if is_export_filesystem_id is None:
